fix: quick ask could not register global shortcuts
